+++ tests/test_resources.py
@@ -113,6 +113,14 @@ class TestResources(unittest.TestCase):
         self.assertEqual(project.name, 'Foo')
         self.assertEqual(project.description, 'Bar')
 
+    def test_supports_setting_of_date_datetime_attributes(self):
+        from datetime import date, datetime
+        issue = self.redmine.issue.new()
+        issue.start_date = date(2014, 3, 9)
+        self.assertEqual(issue.attributes['start_date'], '2014-03-09')
+        issue.start_date = datetime(2014, 3, 9, 20, 2, 2)
+        self.assertEqual(issue.attributes['start_date'], '2014-03-09T20:02:02Z')
+
     def test_supports_setting_of_attributes_via_dict(self):
         project = self.redmine.project.new()
         project['name'] = 'Foo'